Moved redundant state to higher level.
